Moving code to sgpr.py
