--- GPflow/gprfitc.py
@@ -1,51 +1,0 @@
-import tensorflow as tf
-import numpy as np
-from .model import GPModel
-from .param import Param
-from .densities import multivariate_normal
-from .mean_functions import Zero
-from . import likelihoods
-from .tf_hacks import eye
-
-class GPRFITC(GPModel):
-
-    def __init__(self, X, Y, kern, Z, mean_function=Zero()):
-        """
-        This implements GP regression with the FITC approximation. The key reference is
-
-        @INPROCEEDINGS{Snelson06sparsegaussian,
-        author = {Edward Snelson and Zoubin Ghahramani},
-        title = {Sparse Gaussian Processes using Pseudo-inputs},
-        booktitle = {ADVANCES IN NEURAL INFORMATION PROCESSING SYSTEMS },
-        year = {2006},
-        pages = {1257--1264},
-        publisher = {MIT press}
-        }
-
-        X is a data matrix, size N x D
-        Y is a data matrix, size N x R
-        Z is a matrix of pseudo inputs, size M x D
-        kern, mean_function are appropriate GPflow objects
-
-        This method only works with a Gaussian likelihood.
-
-        """
-        likelihood = likelihoods.Gaussian()
-        GPModel.__init__(self, X, Y, kern, likelihood, mean_function)
-        self.Z = Param(Z)
-        self.num_data = X.shape[0]
-        self.num_latent = Y.shape[1]
-
-    def build_likelihood(self):
-        """
-        Constuct a tensorflow function to compute the bound on the marginal
-        likelihood.. 
-        """
-        pass
-        
-    def build_predict(self, Xnew, full_cov=False):
-        """
-        Compute the mean and variance of the latent function at some new points
-        Xnew.
-        """
-        pass
